신고 제재 로직 검토 완료

+++ src/main/java/com/kh/fantimate/mypage1/controller/MypageUserController.java
@@ -133,10 +133,12 @@ public class MypageUserController {
 						String fileName = userAtt.getAttClName();
 						File deleteFile = new File(savePath + "\\" + fileName);
 
-						// 기존 파일 삭제
-						if(deleteFile.exists()) {
-							deleteFile.delete();
-							System.out.println("기존 프로필 삭제 완료");
+						if(!(fileName.equals("profile1.svg") && fileName.equals("profile2.svg") && fileName.equals("profile3.svg"))) {
+							// 기존 파일 삭제
+							if(deleteFile.exists()) {
+								deleteFile.delete();
+								System.out.println("기존 프로필 삭제 완료");
+							}
 						}
 					}
 					// 혹시 처음 업로드라면 경로생성
